Fixed spacing

+++ src/com/egeniq/utils/api/APIUtils.java
@@ -34,12 +34,12 @@ public class APIUtils {
         }
     }
 
-    public static long getLong(JSONObject object, String key, int fallback) {
-    	try {
-    		return object.isNull(key) ? fallback : object.getLong(key);
-    	} catch (JSONException e) {
-    		return fallback;
-    	}
+    public static long getLong(JSONObject object, String key, long fallback) {
+        try {
+            return object.isNull(key) ? fallback : object.getLong(key);
+        } catch (JSONException e) {
+            return fallback;
+        }
     }
     
     public static double getDouble(JSONObject object, String key, double fallback) {
